requirements.txt
tensorflow>=2.3.1
numpy>=1.18.5
keras>=2.4.3
streamlit>=1.27.1
opencv-python>=4.7.0.68
matplotlib>=3.3.2
scikit-learn>=0.9
flask>=1.1.2
Pillow>=8.4.0